Rename AnnotatedTree to Tree
