+++ brushfire-core/src/main/scala/com/stripe/brushfire/Voter.scala
@@ -26,7 +26,7 @@ trait Voter[T, P] { self =>
       f(self.combine(targets))
   }
 
-  final def predict[K, V, A](trees: Iterable[AnnotatedTree[K, V, T, A]], row: Map[K, V])(implicit traversal: TreeTraversal[K, V, T, A], semigroup: Semigroup[T]): P =
+  final def predict[K, V, A](trees: Iterable[Tree[K, V, T, A]], row: Map[K, V])(implicit traversal: TreeTraversal[K, V, T, A], semigroup: Semigroup[T]): P =
     combine(trees.flatMap(_.targetFor(row)))
 }
