+++ brushfire-core/src/main/scala/com/stripe/brushfire/package.scala
@@ -1,6 +1,5 @@
 package com.stripe
 
 package object brushfire {
-  type Tree[K, V, T] = AnnotatedTree[K, V, T, Unit]
 }
