+++ brushfire-finatra/src/main/scala/com/stripe/brushfire/finatra/BrushfireServer.scala
@@ -7,7 +7,7 @@ import com.twitter.finagle.http._
 import com.twitter.bijection._
 
 class BrushfireServer extends FinatraServer {
-  def score[K, V, T: Semigroup, P](root: String, trees: Iterable[Tree[K, V, T]], voter: Voter[T, P])(fn: ParamMap => Map[K, V]) = {
+  def score[K, V, T: Semigroup, P](root: String, trees: Iterable[Tree[K, V, T, Unit]], voter: Voter[T, P])(fn: ParamMap => Map[K, V]) = {
     register(new Controller {
       get(root) { request =>
         val row = fn(request.params)
@@ -17,7 +17,7 @@ class BrushfireServer extends FinatraServer {
     })
   }
 
-  def loadAndScore[K, V, T: Semigroup, P](root: String, treePath: String, voter: Voter[T, P])(fn: ParamMap => Map[K, V])(implicit inj: Injection[Tree[K, V, T], String]) = {
+  def loadAndScore[K, V, T: Semigroup, P](root: String, treePath: String, voter: Voter[T, P])(fn: ParamMap => Map[K, V])(implicit inj: Injection[Tree[K, V, T, Unit], String]) = {
     val trees = scala.io.Source.fromFile(treePath).getLines.map { line =>
       val parts = line.split("\t")
       inj.invert(parts(1)).get
